- Resolved conflicts in setu-validator/src/lib.rs
- Removed old network_service.rs (refactored to network/ module)
- Updated user_handler.rs to work with new architecture
- Integrated new Transfer API from types/transfer.rs

+++ setu-rpc/src/messages.rs
@@ -208,8 +208,8 @@ pub struct SubmitTransferRequest {
     pub from: String,
     /// Receiver address
     pub to: String,
-    /// Amount to transfer
-    pub amount: i128,
+    /// Amount to transfer (non-negative)
+    pub amount: u64,
     /// Transfer type (flux, instant, etc.)
     pub transfer_type: String,
     /// Optional preferred solver
